Improve parser function grammar yet again
