+++ preprocessor/src/main/antlr/com/lucaskjaerozhang/wikitext_parser/grammar/preprocess/WikiTextPreprocessor.g4
@@ -55,9 +55,7 @@ templateParameterParameterValues
    : link
    | template
    | parserFunction
-   | ' '
-   | '"'
-   | '\''
+   | nonControlPunctuation
    | EQUALS
    | TEXT
    | COLON
@@ -117,13 +115,18 @@ parserFunctionParameterValues
    | SLASH
    | OPEN_CARAT
    | CLOSE_CARAT
-   | ' '
-   | '"'
-   | '\''
-   | ';'
+   | nonControlPunctuation
    | unresolvedTemplateParameter
    | parserFunction
    | template
+   ;
+
+nonControlPunctuation
+   : ' '
+   | '.'
+   | '"'
+   | '\''
+   | ';'
    ;
 
 any
