+++ preprocessor/src/test/java/com/lucaskjaerozhang/wikitext_parser/preprocess/function/UrlFunctionEvaluatorTest.java
@@ -1,19 +1,9 @@
 package com.lucaskjaerozhang.wikitext_parser.preprocess.function;
 
-import com.lucaskjaerozhang.wikitext_parser.preprocess.Preprocessor;
-import com.lucaskjaerozhang.wikitext_parser.preprocess.template.provider.DummyTemplateProvider;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class UrlFunctionEvaluatorTest {
-  public static void testParserFunction(String input, String expected) {
-
-    Preprocessor preprocessor =
-        Preprocessor.builder().templateProvider(new DummyTemplateProvider()).build();
-    String result = preprocessor.preprocess(input, true);
-    Assertions.assertEquals(expected, result);
-  }
-
+class UrlFunctionEvaluatorTest extends BaseParserFunctionTest {
   @Test
   void testAnchorEncode() {
     testParserFunction("{{anchorencode:x y z á é}}", "x_y_z_á_é");
